+++ src/external.cpp
@@ -474,6 +474,45 @@ void External::force_backtrack (size_t new_level) {
 
 /*------------------------------------------------------------------------*/
 
+int External::propagate_assumptions () {
+  int res = internal->propagate_assumptions ();
+  if (res == 10 && !extended)
+    extend (); // Call solution reconstruction
+  check_solve_result (res);
+  reset_limits ();
+  return res;
+}
+
+void External::get_entrailed_literals (std::vector<int> &trailed) {
+  std::vector<int> ilit_implicants;
+  internal->get_entrailed_literals (ilit_implicants);
+
+  // Those implied literals must be filtered out that are witnesses
+  // on the reconstruction stack -> no inplace externalize is possible.
+  // (Internal does not see these marks, so no earlier filter is
+  // possible.)
+
+  for (const auto &ilit : ilit_implicants) {
+    assert (ilit);
+    const int elit = internal->externalize (ilit);
+    if (!marked (tainted, elit)) {
+      trailed.push_back (elit);
+    }
+  }
+}
+
+void External::conclude_unknown () {
+  if (!internal->proof || concluded)
+    return;
+  concluded = true;
+
+  vector<int> trail;
+  get_entrailed_literals (trail);
+  internal->proof->conclude_unknown (trail);
+}
+
+/*------------------------------------------------------------------------*/
+
 // Internal checker if 'solve' claims the formula to be satisfiable.
 
 void External::check_satisfiable () {