+++ src/internal.hpp
@@ -1140,6 +1140,11 @@ struct Internal {
   failed_constraint ();     // Was constraint used to proof unsatisfiablity?
   void reset_constraint (); // Reset after 'solve' call.
 
+  // Propagate the current set of assumptions and return the
+  // non-witness assigned literals
+  int propagate_assumptions ();
+  void get_entrailed_literals (std::vector<int> &entrailed);
+
   // Forcing decision variables to a certain phase.
   //
   void phase (int lit);