+++ src/solver.cpp
@@ -691,6 +691,35 @@ void Solver::reset_constraint () {
 
 /*------------------------------------------------------------------------*/
 
+int Solver::propagate () {
+  TRACE ("propagate_assumptions");
+  REQUIRE_VALID_STATE ();
+  transition_to_steady_state ();
+  const int res = external->propagate_assumptions ();
+  if (tracing_nb_lidrup_env_var_method)
+    flush_proof_trace (true);
+  LOG_API_CALL_RETURNS ("propagate_assumptions", res);
+  if (res == 10)
+    STATE (SATISFIED);
+  else if (res == 20)
+    STATE (UNSATISFIED);
+  else
+    STATE (STEADY);
+  return res;
+}
+
+void Solver::get_entrailed_literals (std::vector<int> &entrailed) {
+  TRACE ("get_entrailed_literals");
+  REQUIRE_STEADY_STATE ();
+  external->conclude_unknown ();
+  external->get_entrailed_literals (entrailed);
+  if (tracing_nb_lidrup_env_var_method)
+    flush_proof_trace (true);
+  LOG_API_CALL_RETURNS ("get_entrailed_literals", (int) entrailed.size ());
+}
+
+/*------------------------------------------------------------------------*/
+
 int Solver::call_external_solve_and_check_results (bool preprocess_only) {
   transition_to_steady_state ();
   assert (state () & READY);
@@ -1198,13 +1227,17 @@ bool Solver::disconnect_proof_tracer (FileTracer *tracer) {
 void Solver::conclude () {
   TRACE ("conclude");
   REQUIRE_VALID_STATE ();
-  REQUIRE (state () == UNSATISFIED || state () == SATISFIED,
-           "can only conclude in satisfied or unsatisfied state");
+  REQUIRE (state () == UNSATISFIED || state () == SATISFIED ||
+               state () == STEADY,
+           "can only conclude in satisfied, unsatisfied or STEADY state");
   if (state () == UNSATISFIED)
     internal->conclude_unsat ();
   else if (state () == SATISFIED)
     external->conclude_sat ();
-  assert (state () == UNSATISFIED || state () == SATISFIED);
+  else if (state () == STEADY)
+    external->conclude_unknown ();
+  assert (state () == UNSATISFIED || state () == SATISFIED ||
+          state () == STEADY);
   LOG_API_CALL_END ("conclude");
 }