+++ test/api/parcompwrite.cpp
@@ -12,13 +12,13 @@
 extern "C" {
 #include <pthread.h>
 #include <unistd.h>
-};
+}
 
 using namespace std;
 using namespace CaDiCaL;
 
 static string prefix (const char *tester) {
-  string res = "/tmp/parcompwrite-";
+  string res = "/tmp/cadical-api-test-parcompwrite-";
   res += tester;
   res += "-";
   res += to_string (getpid ());